Add --skip-incompatible flag
This flag makes it possible to skip requirements that are incompatible
with the current environment, i.e. those whose PEP-496 environment
markers do not evaluate thruthly.

For example, when running Python 3.5, previously there would be no way
of processing the following requirement properly:

```
enum34; python_version<"3.4"
```

Ignoring it via `-r` is a bit heavy handed, as it means that it would no
longer get checked when running e.g. Python 2.7.

By using `-s` or `--skip-incompatible`, this requirement gets ignored
when running python versions 3.4 and above, but evaluated when running
on earlier versions.