Fix --skip-incompatible flag
Why it is a bug
---------------

The flag also made the the code skip all other extra requirements. This
was due to a wrongfully added `else` clause, which should not be there:
because of it, the code that added requirements was triggered only `if
not options.skip_incompatible`, which is obviously wrong.

Why this patch fixes it
-----------------------

This patch makes the `explicit.add` code run even when the
`--skip-incompatible` flag is present.

With this patch, both of the branches that ignored the requirements
still work: the `options.ignore_reqs` still uses continue to bail on a
requirement, so does `options.skip_incompatible`.