Item location relative to package, not 'here'
Previously, find_extra_reqs would concatenate the package_location with the file_location for a library import. This fails when using a virtualenv, because the package_location and the file_location end up sharing paths. This PR changes the file_location to be relative to the package, so that you get an actual file path. 

Fixes #75.