ImportVisitor: skip all relative imports
This commit is a step towards detecting when a package is not installed.
Currently, `pip-check-rqs` will silently skip them, even if they are
actually extra/missing in the project requirements.

---

Checking just `node.module` is not sufficient for detecting a relative
imports. It only finds relative imports of the form:

```
from . import foo
```

It misses imports of the following form:

```
from .foo import bar
from .....more_dots import baz
```

... letting the execution pass these to the `_addModule`. This fails
with `ImportError` at `imp.find_module`, because the parameters given to
it to not include the dots from the import statement.

This makes the `ImportError` less useful to us, if we would like to use
it to detect requirements that aren't installed. Right now,
`pip-missing-reqs` will silently skip missing installs, which
contributes to significant user confusion.